backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI RAG backend for Klippy"
dependencies = [
    "fastapi",
    "uvicorn",
    "llama-index",
    "llama-index-vector-stores-qdrant",
    "llama-index-embeddings-openai",
    "llama-index-llms-openai",
    "llama-index-callbacks-arize-phoenix",
    "qdrant-client",
    "redis",
    "python-dotenv",
    "pydantic",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock>=3.15.1",
    "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
